[NS-5] Add deleteUser method to the controller layer and tests for it

+++ src/main/java/com/nerdysoft/library/service/impl/UserServiceImpl.java
@@ -2,6 +2,7 @@ package com.nerdysoft.library.service.impl;
 
 import static com.nerdysoft.library.exceptionhandler.ExceptionMessages.MAX_USER_BOOKS_QUANTITY;
 import static com.nerdysoft.library.exceptionhandler.ExceptionMessages.USER_BOOK_RELATION_ALREADY_EXISTS;
+import static com.nerdysoft.library.exceptionhandler.ExceptionMessages.USER_BORROWED_BOOKS;
 
 import com.nerdysoft.library.exceptionhandler.ExceptionMessages;
 import com.nerdysoft.library.exceptionhandler.exceptions.BookAmountConflictException;
@@ -33,10 +34,22 @@ public class UserServiceImpl implements UserService {
   @Value("${application.max-book-quantity-for-user}")
   private int maxBookQuantityForUser;
 
+  /**
+   * Deletes a user that has no borrowed books.
+   *
+   * @param userId - a user ID
+   */
   @Override
   @Transactional
   public void deleteUser(UUID userId) {
     User user = findUserById(userId);
+    int userBookQuantity = userRepository.countBookRelationsByUserId(userId.toString());
+
+    if (userBookQuantity != 0) {
+      log.debug(USER_BORROWED_BOOKS.formatted(userId, userBookQuantity));
+      throw new UserBookRelationConflictException(
+          USER_BORROWED_BOOKS.formatted(userId, userBookQuantity));
+    }
     userRepository.delete(user);
   }
 
@@ -54,13 +67,13 @@ public class UserServiceImpl implements UserService {
     verifyIfUserBookRelationAlreadyExists(userId, bookId);
     verifyIfUserExists(userId);
     BookDto bookDto = bookService.getBookById(bookId);
-    int userBooksQuantity = userRepository.countBookRelationsById(userId.toString());
+    int userBooksQuantity = userRepository.countBookRelationsByUserId(userId.toString());
 
     if (bookDto.getAmount() == 0) {
       log.debug(ExceptionMessages.NO_BOOKS);
       throw new BookAmountConflictException(ExceptionMessages.NO_BOOKS);
     } else if (userBooksQuantity >= maxBookQuantityForUser) {
-      log.debug(MAX_USER_BOOKS_QUANTITY);
+      log.debug(MAX_USER_BOOKS_QUANTITY.formatted(maxBookQuantityForUser));
       throw new UserBookRelationConflictException(
           MAX_USER_BOOKS_QUANTITY.formatted(maxBookQuantityForUser));
     } else {
